refactor(Color): Color는 Pawn만의 것이 아니므로 모두의 것으로 리팩토링
패키지 트리도 변경함
