+++ src/main/java/pieces/Pawn.java
@@ -1,16 +1,16 @@
-package pieces.pawn;
+package pieces;
 
 import lombok.Getter;
 
-import static pieces.pawn.PawnColor.BLACK_PAWN;
-import static pieces.pawn.PawnColor.WHITE_PAWN;
+import static pieces.PieceColor.BLACK_PIECE;
+import static pieces.PieceColor.WHITE_PIECE;
 
 @Getter
 public class Pawn {
     private final String color;
 
     public Pawn() {
-        this.color = WHITE_PAWN;
+        this.color = WHITE_PIECE;
     }
 
     public Pawn(String color) {
@@ -23,7 +23,7 @@ public class Pawn {
      * TODO: 던진 예외를 처리해야 함
      */
     private void validatePawnColor(final String color) {
-        if (!color.equals(BLACK_PAWN) && !color.equals(WHITE_PAWN)) {
+        if (!color.equals(BLACK_PIECE) && !color.equals(WHITE_PIECE)) {
             throw new IllegalArgumentException("[ERROR] black 또는 white 폰을 입력해야 합니다. color: " + color);
         }
     }
