+++ src/main/java/pieces/PieceColor.java
@@ -1,0 +1,6 @@
+package pieces;
+
+public class PieceColor {
+    public static final String WHITE_PIECE = "white";
+    public static final String BLACK_PIECE = "black";
+}
