--- src/main/java/pieces/pawn/PawnColor.java
@@ -1,6 +1,0 @@
-package pieces.pawn;
-
-public class PawnColor {
-    public static final String WHITE_PAWN = "white";
-    public static final String BLACK_PAWN = "black";
-}
